requirements-dev.txt
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
pylint>=3.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
requests-mock>=1.10.0
pre-commit>=3.0.0
mypy>=1.0.0
bandit>=1.7.0